renamed beacon files
